refactor(analyzer): move ClaudeWrapper out of shared extract_json into Claude backend
The ClaudeWrapper envelope parsing was hardcoded in the shared
extract_json() function that all backends (Claude, Codex, Gemini) call.
This meant non-Claude backends were unnecessarily attempting to parse
Claude-specific response formats.

Move ClaudeWrapper struct and try_claude_wrapper() into claude.rs where
they belong. Claude's parse_response() now unwraps the envelope first,
then falls back to generic extract_json(). Other backends are unaffected.
