+++ src/analyzer/backend/claude.rs
@@ -4,8 +4,8 @@
 //! for analysis. Optionally uses `--json-schema` for structured output.
 
 use super::{
-    extract_json, parse_rate_limit_info, wait_with_timeout, AgentBackend, BackendError,
-    BackendResult, RawMarker, MARKER_JSON_SCHEMA,
+    extract_json, extract_json_inner, parse_rate_limit_info, wait_with_timeout, AgentBackend,
+    AnalysisResponse, BackendError, BackendResult, RawMarker, MARKER_JSON_SCHEMA,
 };
 use crate::analyzer::TokenBudget;
 use serde::Deserialize;
@@ -120,6 +120,11 @@ impl AgentBackend for ClaudeBackend {
     }
 
     fn parse_response(&self, response: &str) -> BackendResult<Vec<RawMarker>> {
+        // Try Claude CLI wrapper envelope first
+        if let Some(result) = try_claude_wrapper(response.trim()) {
+            return result.map(|a| a.markers);
+        }
+        // Fall back to generic JSON extraction
         let analysis = extract_json(response)?;
         Ok(analysis.markers)
     }
@@ -134,6 +139,56 @@ impl AgentBackend for ClaudeBackend {
 struct ClaudeErrorWrapper {
     is_error: Option<bool>,
     result: Option<String>,
+}
+
+/// Claude CLI wrapper envelope format.
+///
+/// When invoked with `--output-format json`, the Claude CLI wraps the actual
+/// response inside this envelope. The inner `result` contains the model's text
+/// output, and `structured_output` (when using `--json-schema`) contains the
+/// parsed JSON directly.
+#[derive(Debug, Deserialize)]
+struct ClaudeWrapper {
+    #[serde(rename = "type")]
+    response_type: Option<String>,
+    is_error: Option<bool>,
+    result: Option<String>,
+    structured_output: Option<AnalysisResponse>,
+}
+
+/// Try to unwrap a Claude CLI wrapper envelope.
+///
+/// Returns `Some(Ok(response))` if the wrapper was successfully unwrapped,
+/// `Some(Err(..))` if the wrapper indicates an error, or `None` if the
+/// input is not a Claude wrapper (allowing the caller to fall through to
+/// generic extraction).
+fn try_claude_wrapper(response: &str) -> Option<BackendResult<AnalysisResponse>> {
+    let wrapper: ClaudeWrapper = serde_json::from_str(response).ok()?;
+
+    // Must have type == "result" to be a Claude wrapper
+    if wrapper.response_type.as_deref() != Some("result") {
+        return None;
+    }
+
+    // Check for error responses
+    if wrapper.is_error == Some(true) {
+        return Some(Err(BackendError::JsonExtraction {
+            response: wrapper.result.unwrap_or_else(|| "Claude error".to_string()),
+        }));
+    }
+
+    // Try structured_output first (from --json-schema)
+    if let Some(analysis) = wrapper.structured_output {
+        return Some(Ok(analysis));
+    }
+
+    // Try to parse the inner result string
+    let inner = wrapper.result.as_deref().unwrap_or("");
+    if inner.is_empty() {
+        return None;
+    }
+
+    Some(extract_json_inner(inner))
 }
 
 /// Extract error message from Claude's JSON response wrapper.
@@ -151,7 +206,12 @@ fn extract_error_from_claude_response(stdout: &str) -> Option<String> {
 
 #[cfg(test)]
 mod tests {
+    use super::super::MarkerCategory;
     use super::*;
+
+    // ============================================
+    // ClaudeBackend basic tests
+    // ============================================
 
     #[test]
     fn claude_backend_name() {
@@ -196,5 +256,98 @@ mod tests {
 
         let result = backend.parse_response(response);
         assert!(result.is_err());
+    }
+
+    // ============================================
+    // try_claude_wrapper unit tests
+    // ============================================
+
+    #[test]
+    fn try_claude_wrapper_non_wrapper_json() {
+        // Valid JSON but type != "result" => returns None (fall through)
+        let json = r#"{"type":"other","result":"foo","is_error":false}"#;
+        let result = try_claude_wrapper(json);
+        assert!(result.is_none());
+    }
+
+    #[test]
+    fn try_claude_wrapper_non_json() {
+        // Not valid JSON at all => returns None
+        let result = try_claude_wrapper("this is plain text");
+        assert!(result.is_none());
+    }
+
+    #[test]
+    fn try_claude_wrapper_error_wrapper() {
+        // is_error: true => Some(Err(JsonExtraction))
+        let json = r#"{"type":"result","is_error":true,"result":"Something went wrong"}"#;
+        let result = try_claude_wrapper(json).expect("should return Some");
+        assert!(matches!(result, Err(BackendError::JsonExtraction { .. })));
+    }
+
+    #[test]
+    fn try_claude_wrapper_empty_result_returns_none() {
+        // type=result, is_error=false, result="" and no structured_output => None
+        let json = r#"{"type":"result","is_error":false,"result":""}"#;
+        let result = try_claude_wrapper(json);
+        assert!(result.is_none());
+    }
+
+    #[test]
+    fn try_claude_wrapper_missing_type_returns_none() {
+        // No "type" field => response_type is None => not "result" => returns None
+        let json = r#"{"is_error":false,"result":"{\"markers\":[]}"}"#;
+        let result = try_claude_wrapper(json);
+        assert!(result.is_none());
+    }
+
+    // ============================================
+    // Claude wrapper through parse_response tests
+    // ============================================
+
+    #[test]
+    fn parse_response_wrapper_with_code_block() {
+        let backend = ClaudeBackend::new();
+        let response = r#"{"type":"result","subtype":"success","is_error":false,"result":"```json\n{\"markers\":[{\"timestamp\":10.0,\"label\":\"Test\",\"category\":\"success\"}]}\n```"}"#;
+        let markers = backend.parse_response(response).unwrap();
+        assert_eq!(markers.len(), 1);
+        assert_eq!(markers[0].category, MarkerCategory::Success);
+    }
+
+    #[test]
+    fn parse_response_wrapper_direct_json() {
+        let backend = ClaudeBackend::new();
+        let response = r#"{"type":"result","is_error":false,"result":"{\"markers\":[{\"timestamp\":5.0,\"label\":\"Plan\",\"category\":\"planning\"}]}"}"#;
+        let markers = backend.parse_response(response).unwrap();
+        assert_eq!(markers.len(), 1);
+        assert_eq!(markers[0].category, MarkerCategory::Planning);
+    }
+
+    #[test]
+    fn parse_response_wrapper_structured_output() {
+        let backend = ClaudeBackend::new();
+        let response = r#"{"type":"result","subtype":"success","is_error":false,"result":"","structured_output":{"markers":[{"timestamp":10.0,"label":"Schema output","category":"success"}]}}"#;
+        let markers = backend.parse_response(response).unwrap();
+        assert_eq!(markers.len(), 1);
+        assert_eq!(markers[0].label, "Schema output");
+        assert_eq!(markers[0].category, MarkerCategory::Success);
+    }
+
+    #[test]
+    fn parse_response_wrapper_error() {
+        let backend = ClaudeBackend::new();
+        let response =
+            r#"{"type":"result","is_error":true,"result":"Failed to analyze: content too large"}"#;
+        let result = backend.parse_response(response);
+        assert!(matches!(result, Err(BackendError::JsonExtraction { .. })));
+    }
+
+    #[test]
+    fn parse_response_wrapper_empty_markers() {
+        let backend = ClaudeBackend::new();
+        let response =
+            r#"{"type":"result","is_error":false,"result":"```json\n{\"markers\":[]}\n```"}"#;
+        let markers = backend.parse_response(response).unwrap();
+        assert!(markers.is_empty());
     }
 }
